pyproject.toml
[project]
name = "petmed-backend"
version = "0.1.0"
description = "AI-Powered Veterinary Diagnosis & Treatment Support System"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "PetMed Team" }]
dependencies = [
    # Core
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # AWS
    "boto3>=1.35.0",
    "aioboto3>=13.0.0",
    "python-jose[cryptography]>=3.3.0",

    # Google AI
    "google-genai>=1.8.0",
    "google-cloud-aiplatform>=1.70.0",

    # Vector DB
    "pinecone[grpc]>=5.0.0",

    # Task Queue
    "redis>=5.0.0",
    "arq>=0.27.0",

    # PDF Processing
    "unstructured[pdf]>=0.16.0",
    "pypdf>=4.0.0",
    "PyMuPDF>=1.24.0",

    # Logging
    "structlog>=24.4.0",
    "asgi-correlation-id>=4.0.0",

    # Utilities
    "python-ulid>=3.0.0",
    "httpx>=0.27.0",
    "orjson>=3.10.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "moto[dynamodb,s3,cognitoidp]>=5.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "boto3-stubs[dynamodb,s3,cognito-idp]>=1.35.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
